ai.hohyub.site/rag/llama/app/requirements.txt
# LangChain packages
langchain-core
langchain-community
langchain-postgres
langchain-huggingface

# Database
sqlalchemy>=2
psycopg[binary]
asyncpg
pgvector

# FastAPI
fastapi
uvicorn[standard]
python-dotenv
pydantic

# Local LLM & Embeddings (CPU only - torch will be installed separately)
transformers>=4.36.0
accelerate
safetensors
sentence-transformers
peft
datasets
trl>=0.7.0
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
bitsandbytes-windows

